Restructure nuclear_dyn: flags -> inputfile params
Since command-line flags given to nuclear_dyn.py are not included in the
input files, the calculations could not be reproduced if one is given
only the files. Therefore, the argument structure is reorganized:
If pre-calculated "Franck-Condon overlap integrals" (which may or may
not include weighting functions from the Gammma(R) dependence), only
partial incorporation of the Gamma(R) dependence or restricting the
calculation to resonance-state projections (for the wavepacket dynamics)
are desired, new parameters in the input file must be set accordingly.
Only the names of additional input files can be given as command-line
flags (-f, -g, -F; see documentation of nuclear_dyn.py).
Other programs were changed to be able to work with this changed
structure.